+++ tpa/apirest/src/main/java/com/audiolibrary/rpa/apirest/controller/ArtistController.java
@@ -10,6 +10,9 @@ import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
+
 @RestController
 @RequestMapping(value = "/artists")
 public class ArtistController {
@@ -17,18 +20,20 @@ public class ArtistController {
     @Autowired
     private ArtistRepository artistRepository;
 
-    @Autowired
-    private AlbumRepository albumRepository;
 
     // GET /artists/1
     @RequestMapping(method = RequestMethod.GET,
             produces = "application/json",
             value = "/{id}")
-    public Artist findById(@PathVariable(value = "id") Integer id){
+    public Artist findById(@PathVariable(value = "id") Integer id) throws EntityNotFoundException {
         Artist artist = artistRepository.findOne(id);
+        if(artist == null){
+            throw new EntityNotFoundException("L'artiste d'identifiant " + id + " n'a pas été trouvé");
+        }
         return artist;
     }
 
+    // GET /artists?name=aerosmith
     @RequestMapping(value = "", params = {"name", "page", "size", "sortDirection", "sortProperty"})
     public Page<Artist> findByName(
             @RequestParam("name") String name,
@@ -41,6 +46,7 @@ public class ArtistController {
     }
 
 
+        // GET /artists?page=0&size=10&sortProperty=name&sortDirection=ASC
         @RequestMapping(method = RequestMethod.GET,
         produces = "Application/json",
         value="", params = {"page", "size", "sortDirection", "sortProperty"})
@@ -53,29 +59,43 @@ public class ArtistController {
         return artistRepository.findAll(pageRequest);
         }
 
-
+    // 4 - Création d'un artiste
+    // POST /artists avec les données de l'artiste en JSON dans le champ data de la requête.
     @RequestMapping(method = RequestMethod.POST,
-    consumes = "Application/json",
-    value="")
-    public Artist createArtist(@RequestBody Artist artist) {
+            consumes = "Application/json",
+            value = "")
+    public Artist createArtist(@RequestBody Artist artist) throws EntityExistsException{
         String name = artist.getName();
+
+        if(!artistRepository.findByName(name).isEmpty()) {
+            throw new EntityExistsException("L'artiste nommé " + name + " existe déjà, oui oui, vous ne pouvez pas l'ajouter à nouveau !");
+        }
         return artistRepository.save(artist);
     }
 
+    // PUT /artists/4 avec les données de l'artiste en JSON dans le champ data de la requête
     @RequestMapping(method = RequestMethod.PUT,
-    consumes = "Application/json",
-    value = "/{id}")
-    public Artist editArtist(@PathVariable Integer id, @RequestBody Artist artist) {
-        String name = artist.getName();
+            consumes = "Application/json",
+            value = "/{id}")
+    public Artist editArtist(@PathVariable Integer id, @RequestBody Artist artist) throws EntityNotFoundException {
+        Artist a = artistRepository.findOne(id);
+        String name = a.getName();
+        if(artistRepository.findByName(name).isEmpty()){
+            throw new EntityNotFoundException("L'artiste " + name +" que vous souhaitez modifier n'existe pas");
+        }
         return artistRepository.save(artist);
     }
 
+    // DELETE /artists/5
     @RequestMapping(method = RequestMethod.DELETE,
-    value = "/{id}")
+            value = "/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteArtist(@PathVariable Integer id) {
-        Artist artist = artistRepository.findOne(id);
-        String name = artist.getName();
+    public void deleteArtist(@PathVariable Integer id) throws EntityNotFoundException {
+        Artist a = artistRepository.findOne(id);
+        String name = a.getName();
+        if(artistRepository.findByName(name).isEmpty()){
+            throw new EntityNotFoundException("L'artiste " + name + " n'existe pas, vous ne pouvez pas le supprimer");
+        }
         artistRepository.delete(id);
     }
